% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{convert_ternaries_simple}
\alias{convert_ternaries_simple}
\title{Convert ternary operators to CppAD::CondExp* - Simplified robust version}
\usage{
convert_ternaries_simple(txt)
}
\arguments{
\item{txt}{Character string containing C++ code with ternary operators}
}
\value{
Character string with ternaries converted to CppAD::CondExp*
}
\description{
This function uses a straightforward iterative approach to convert ternary
operators. It processes them from innermost to outermost.
}
