% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ensurePythonEnv}
\alias{ensurePythonEnv}
\title{Ensure availability of a dedicated Python environment with required packages}
\usage{
ensurePythonEnv(
  envname = "CppODE",
  packages = c("numpy", "sympy"),
  verbose = FALSE
)
}
\arguments{
\item{envname}{Character scalar. Name of the Python virtual environment
to check or create. Default is \code{"CppODE"}.}

\item{packages}{Character vector. Python packages to ensure are installed.
Default: c("numpy", "sympy").}

\item{verbose}{Logical. If TRUE, prints progress messages.}
}
\value{
Invisibly returns the name of the environment that was ensured.
}
\description{
This helper function checks whether a Python virtual environment with a
given name (default: \code{"CppODE"}) exists and contains the required
Python packages (default: c("numpy", "sympy")). If not, it will create
the virtual environment and install the missing packages.
}
\details{
After ensuring availability, the environment is activated via
\code{reticulate::use_virtualenv()}.
}
