% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{find_condition_start_simple}
\alias{find_condition_start_simple}
\title{Find start of condition for a ternary operator}
\usage{
find_condition_start_simple(txt, question_pos)
}
\description{
Find start of condition for a ternary operator
}
