% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{find_matching_colon_simple}
\alias{find_matching_colon_simple}
\title{Find matching colon for a question mark}
\usage{
find_matching_colon_simple(txt, question_pos)
}
\description{
Find matching colon for a question mark
}
