% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppODE.R
\name{CppFun}
\alias{CppFun}
\title{Generate full C++ code for an ODE model with sensitivities}
\usage{
CppFun(
  odes,
  events = NULL,
  modelname = NULL,
  deriv = TRUE,
  secderiv = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{odes}{Named character vector of ODEs.}

\item{events}{Optional data.frame of fixed-time events (see GetBoostObserver).}

\item{modelname}{Optional string, defaults to random.}

\item{deriv}{Logical, include first derivatives (Jacobian)?}

\item{secderiv}{Logical, include second derivatives (Hessian)?}

\item{verbose}{Logical, print progress?}

\item{fixed}{Not yet implemented.}

\item{compile}{Not yet implemented.}
}
\value{
Character name of file
}
\description{
This function orchestrates ODE definition, Jacobian generation,
event handling, and sensitivity calculation into a complete C++ file.
}
