% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{getSymbols}
\alias{getSymbols}
\title{Extract symbols from expressions}
\usage{
getSymbols(char, exclude = NULL)
}
\arguments{
\item{char}{Character vector of expressions.}

\item{exclude}{Optional character vector of names to exclude.}
}
\value{
Character vector with unique symbol names.
}
\description{
Parses a character vector of expressions and returns all symbol names
(variable identifiers).
}
\examples{
getSymbols("a*b + c")
getSymbols(c("x + y", "z"), exclude = "y")

}
\author{
Daniel Kaschek
}
\keyword{internal}
