% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{Sympy2CppCode}
\alias{Sympy2CppCode}
\title{Convert a SymPy expression to C++ code suitable for AD solvers}
\usage{
Sympy2CppCode(expr, states, params, n, time_var, expr_name = NULL)
}
\arguments{
\item{expr}{SymPy expression.}

\item{states}{Character vector of state variable names.}

\item{params}{Character vector of parameter names.}

\item{n}{Number of states.}

\item{time_var}{Name of the time variable.}

\item{expr_name}{Optional. Name of the ODE for error messages.}
}
\description{
This internal function converts a SymPy expression into C++ code, replacing:
\itemize{
\item states with \verb{x[0..n-1]},
\item initial values with \verb{params[0..n-1]},
\item parameters with \code{params[n..]},
\item time symbol with \code{t}.
}
}
\details{
Throws an error if the expression contains functions not supported in C++.
}
\author{
Simon Beyer, \email{simon.beyer@fdm.uni-freiburg.de}
}
